Extract snackbar
