+++ mikado-app/src/pages/Graph/Plaza.css
@@ -1,0 +1,5 @@
+
+main {
+  height: 100%;
+  width: 100%;
+}
